+++ CHANGES.md
@@ -5,10 +5,12 @@ Ongoing (for Beta 4)
 ---------------------
 * added onFocus()
 * added per()
+* added reduce()
 * on() will set 'this' to a Minified list containing the event source, instead of the event source itself.
 * $.request()'s headers, username and password parameters have been moved into an object as 4th argument. 
   You can now also set XMLHttpRequest properties.
 * onOver()'s callback does not support the index argument anymore.
+* all iterator function will now get the object they are iterating over in 'this'
 
 
 Beta 3.1 (2013-11-13)